BAEL-1706 move code to guest module
