+++ guest/core-java/pom.xml
@@ -41,11 +41,18 @@
             <version>${org.hamcrest.version}</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.assertj</groupId>
+            <artifactId>assertj-core</artifactId>
+            <version>${assertj.version}</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <properties>
         <log4j2.version>2.8.2</log4j2.version>
         <org.hamcrest.version>1.3</org.hamcrest.version>
+        <assertj.version>3.6.1</assertj.version>
     </properties>
 
 </project>
