+++ guest/core-java/src/test/java/com/stackify/optionalparams/OptionalParamsUnitTest.java
@@ -1,4 +1,4 @@
-package com.baeldung.optionalparams;
+package com.stackify.optionalparams;
 
 import static org.assertj.core.api.Assertions.assertThat;
